Added integration tests

+++ src/main/java/dev/nicklasw/messageboard/domain/message/service/MessageServiceImpl.java
@@ -3,9 +3,12 @@ package dev.nicklasw.messageboard.domain.message.service;
 import java.util.Optional;
 
 import com.querydsl.core.types.Predicate;
-import dev.nicklasw.messageboard.adapter.driven.persistence.MessageRepository;
-import dev.nicklasw.messageboard.adapter.driver.api.exception.MissingEntityException;
+import dev.nicklasw.messageboard.adapter.driven.persistence.message.MessageRepository;
+import dev.nicklasw.messageboard.adapter.driven.security.AuthenticationService;
+import dev.nicklasw.messageboard.adapter.driver.api.common.exception.MissingEntityException;
+import dev.nicklasw.messageboard.adapter.driver.api.common.exception.NotSupportedException;
 import dev.nicklasw.messageboard.domain.message.entities.Message;
+import dev.nicklasw.messageboard.domain.user.entities.User;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -18,6 +21,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class MessageServiceImpl implements MessageService {
 
     private final MessageRepository messageRepository;
+    private final AuthenticationService authenticationService;
 
     @Override
     public Page<Message> findAll(@NonNull final Predicate predicate, @NonNull final Pageable pageable) {
@@ -36,19 +40,22 @@ public class MessageServiceImpl implements MessageService {
     }
 
     @Override
+    @Transactional
     public Message create(@NonNull final Message message) {
-        // TODO, validation
-        // TODO, push event
-
         messageRepository.save(message);
 
         return message;
     }
 
     @Override
+    @Transactional
     public Message update(@NonNull final Message message) {
-        // TODO, validation, should only be able to updated it owns messages
-        // TODO, push event
+        final User user = authenticationService.optionalAuthenticatedUser()
+            .orElseThrow();
+
+        if (message.isNotOwner(user)) {
+            throw new NotSupportedException("You can only update your own messages");
+        }
 
         messageRepository.save(message);
 
@@ -67,11 +74,14 @@ public class MessageServiceImpl implements MessageService {
     //    @Retryable(StaleStateException.class) // TODO
     @Transactional
     public Message delete(@NonNull final Message message) {
-        // TODO, validation, should only be able to delete its own messages
+        final User user = authenticationService.optionalAuthenticatedUser()
+            .orElseThrow();
+
+        if (message.isNotOwner(user)) {
+            throw new NotSupportedException("You can only delete your own messages");
+        }
 
         messageRepository.delete(message);
-
-        //        applicationEventPublisher.publishEvent(new SenderEvent(DELETED, sender));
 
         return message;
     }
